circuit2022_PPG-main/requirements.txt
asgiref==3.5.2
async-generator==1.10
attrs==21.4.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
cryptography==37.0.4
Django==4.0.4
django-yamlfield==1.2.0
h11==0.13.0
html5lib==1.1
idna==3.3
importlib-metadata==4.12.0
inflect==5.6.2
jellyfish==0.9.0
joblib==1.1.0
lxml==4.9.1
Markdown==3.3.7
mysql-connector-python-rf==2.2.2
networkx==2.8.4
nltk==3.7
numpy==1.22.4
outcome==1.2.0
pandas==1.4.2
Pillow==9.2.0
psutil==5.9.1
psycopg2-binary==2.9.3
pycparser==2.21
pyOpenSSL==22.0.0
pypi==2.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
regex==2022.6.2
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scikit-learn==1.1.1
scipy==1.8.1
segtok==1.5.11
selenium==4.3.0
six==1.16.0
sklearn==0.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stopwords==1.0.0
tabulate==0.8.10
threadpoolctl==3.1.0
tqdm==4.64.0
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.10
utils==1.0.1
webencodings==0.5.1
wsproto==1.1.0
yake==0.4.8
zipp==3.8.1
